Loader: rename three methods to better reflect their intentions
_example_groups_for -> _top_level_classes_in: at the beginning,
top-level classes are extracted, from either module objects or
class objects themselves (in the case of nested examples). But
they are not example groups yet.

_is_example_group -> _is_name_of_example_group: the parameter
is expected to be a string, hence it's not really a candidate
for an example group. It's a name of something, and the method
determines whether that something is an example group.

_is_hook -> _is_name_of_hook: same as above.