Settings: no need to use properties for the moment, so remove them
There is no additional logic which needs to be executed
on property access or change. And all members of Settings
are being accessed after creation, so they can be considered
public for all purposes. Hence we don't need the property
mechanism, and it can be introduced without breaking anything
whenever it's needed.