nodetransformers: `when` -> `scope_of_hook`
`when` can take values 'all' and 'each', which do indicate *when*
the hook is to be run, but so do 'before' and 'after'. What's
unique to 'all' and 'each' is that they provide the scope of
the hook: whether they affect each example or are run only at the
edges.