nodetransformers: bail out earlier if `with` statement doesn't have the expected structure
The 'relevancy' of a `with` statement is defined as whether it has the structure
of an example group, example or hook declaration. This makes the transformation
bail out early when encountering `with` statements not matching those schemas,
such as the `with patch.object(...)` in 'spec/mock_patch_example_spec.py'.

However, those `with` statements with the structure of an example group, example
or hook declaration but not actually being one of those are not detected.
Examples:

    ```
    with my_context_manager('this is not a spec heading'):
        ...
    ```

    ```
        with my_thing.not_a_hook:
            ...
    ```

That's exactly the reason why the last `return node` is (at least for now)
necessary (since, in these cases, the `name` is not any mamba identifier,
hence control flows out of all the `if`s).