Remove NodeShouldNotBeTransformed exception
Now transformer classes can (and should) be queried for their ability to
transform the node given to them in construction.

The rationale is that, since all with statements are to be processed
indiscriminately, instantiating a transformer which can't actually transform
that particular with statement is not something exceptional but something that
will happen all the time. Hence it should not be signaled using exceptions.